Close documentation node when filters are changed
