+++ supply-chain/src/styles/styles.scss
@@ -311,11 +311,7 @@ a {
   .graph-node-wrapper {
     display: inline;
     position: relative;
-<<<<<<< HEAD
     vertical-align: top;
-=======
-    vertical-align: bottom;
->>>>>>> 6536f1b766fff24eda34d7d499b537d138a7c056
   }
 }
